# Use buffer pools for decompression