While profiling my application, I've noticed that Sarama spends a significant amount of time growing slices when it's decompressing messages:

![growSlice](https://github.com/Shopify/sarama/assets/1172459/40873d99-9f33-4a46-a553-d534b7bc31c5)

This patch uses an intermediate buffer from a  pool for decompression, to avoid these allocations.
When the final size is know, the result buffer can be allocated and copied.

After the patch:

![withpool2](https://github.com/Shopify/sarama/assets/1172459/832f4db5-7b20-4e57-bd0c-3456b8052215)